feat(ui): add DHCP edit form to the network tab
Add the form to edit DHCP snippets in the DHCP table in machine details.

Fixes: #1890.
