+++ ui/src/app/base/components/DhcpForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DhcpForm";
