+++ ui/src/app/base/components/DhcpForm/types.ts
@@ -1,0 +1,10 @@
+import type { DHCPSnippet } from "app/store/dhcpsnippet/types";
+
+export type DHCPFormValues = {
+  description: DHCPSnippet["description"];
+  enabled: DHCPSnippet["enabled"];
+  entity: string;
+  name: DHCPSnippet["name"];
+  value: DHCPSnippet["value"];
+  type: "" | "subnet" | "controller" | "machine" | "device";
+};
